added keyboard events: press, release, and hold

+++ src/GameObject.h
@@ -4,8 +4,6 @@
 #include <iostream>
 #include <vector>
 #include <unordered_map>
-#include <utility>
-#include <set>
 #include "Structs.h"
 
 /* The basic outline of a game object
@@ -36,7 +34,7 @@ class GameObject {
         double acc_ang;
 
     private:
-        uint32_t flags; 
+        uint16_t flags; 
 
     public:
         GameObject(const std::string& name,
@@ -60,12 +58,11 @@ class GameObject {
 
         std::unordered_map<GameEvent, std::vector<ActionList>> event_list;
         //std::unordered_map<string, std::vector<ActionList>> collision_event_list;
-        //std::unordered_map<SDL_Keypress, std::vector<ActionList>> key_event_list;
 
         void addEvent(GameEvent event, GameAction action, const std::string& name, double value);
         void addEvent(GameEvent event, GameAction action, ObjectAttribute att, double value);
+
         //void addCollisionEvent()
-        //void addKeyEvent()
 
         void runEventList();
